Plugins: Change most *dump plugins to --dump
